Driver points and constructor points are now calculated and automated
This new change allows the user to select the driver and assign points which will go into the f1driver table. the constructor points are calculated based on the driver with the same team id adding up the points collected. finally the values will then be put into the main f1 table.

== NEXT STEP ==
adding a combo box to automate the process off assigning the driver point as it differs based on the position they are in

Adding a tick box of some sort to assign the additional point the driver gets for fast lap and adds it on to their points

Visual Redesign of the form

validation where appropriate

ensuring the data from the f1driver and f1team table are being transferred into the f1table on initial start up